requirements.txt
fastapi==0.115.12
playwright==1.52.0
uvicorn==0.34.3
python-dotenv==1.1.0
python-jose==3.5.0
requests==2.32.4
pytz==2025.2
datetime==5.5
Unidecode==1.4.0